requirements.txt
# config
appdirs
pyyaml

# handlers
requests # webhook outputs
systemd-python # systemd input